pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-server-qdrant"
version = "0.1.0"
description = "Minimalistic MCP server for Qdrant vector database with FastMCP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.1.0",
    "qdrant-client>=1.6.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.24.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[project.scripts]
mcp-server-qdrant = "qdrant_server:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
target-version = "py38"
line-length = 88
